Address slow loading of graduates page
Graduates page was taking over 30 seconds to load information. As a result, async functions have been introduced.
As psycopg2-binary did not support async await functions, psycopg3 was installed in the virtual environment and the connection code to the database was adjusted accordingly.

aiohttp was also installed to aid the async functions.

To counter the SSL error, it was mentioned in the aiohttp function and the value changed to false.

pip freeze > requirements.txt to get all the dependencies used for the server